+++ backend/routes/api/session.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const asyncHandler = require('express-async-handler');
 
-const { validateLogin } = require('../utils/sessionValidators');
+const { validateLogin } = require('../utils/validators');
 const { setTokenCookie, restoreUser } = require('../../utils/auth');
 const { User } = require('../../db/models');
